Move widgets' JS into the <head> of widgets.php. Should fix text/rss widget controls. see #4169
git-svn-id: https://develop.svn.wordpress.org/branches/2.2@5343 602fd350-edb4-49c9-b593-d223f7449a82
